server/requirements.txt
amqp==2.1.4
anyjson==0.3.3
appdirs==1.4.3
asgi-redis==1.3.0
asgiref==1.1.1
attrs==16.3.0
autobahn==0.18.2
Automat==0.5.0
billiard==3.3.0.23
celery==3.1.23
channels==1.1.3
constantly==15.1.0
daphne==1.2.0
Django==1.11
django-material==0.13.1
djangorestframework==3.6.2
incremental==16.10.1
kombu==3.0.35
msgpack-python==0.4.8
olefile==0.44
packaging==16.8
Pillow==4.1.1
pyparsing==2.2.0
python-dateutil==2.6.0
pytz==2017.2
redis==2.10.5
six==1.10.0
Twisted==17.1.0
txaio==2.7.0
vine==1.1.3
zope.interface==4.4.0
